pyproject.toml
[project]
name = "ragchat"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "litellm>=1.82.0",
    "pypdf>=6.7.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "respx>=0.22.0",
    "rich>=14.3.3",
    "ruff>=0.15.4",
    "typer>=0.24.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
